20250124 debug MRDicomProcessingStrategy process dicom.save_as

+++ code_ai/dicom2nii/convert/convert_nifti_postprocess.py
@@ -1,6 +1,7 @@
 import argparse
 import pathlib
 import re
+import traceback
 from abc import ABCMeta, abstractmethod
 from concurrent.futures import ThreadPoolExecutor
 
@@ -29,7 +30,6 @@ class ProcessingStrategy(metaclass=ABCMeta):
             swan_pattern = self.pattern.match(series_path.name)
             if swan_pattern:
                 if series_path.stat().st_size < self.FILE_SIZE:
-                    print('del_file', series_path)
                     json_file_path = series_path.parent.joinpath(series_path.name.replace(r'.nii.gz', '.json'))
                     if json_file_path.exists():
                         json_file_path.unlink()
@@ -40,13 +40,16 @@ class ProcessingStrategy(metaclass=ABCMeta):
         pattern_result = pattern.match(series_path.name)
         if pattern_result:
             groups = pattern_result.groups()
-            # print(f'groups {groups}')
             if len(groups[1]) > 0:
-                suffix_char = pattern_result.groups()[1]
-                suffix_int = ord(suffix_char) - self.CHAR_OFFSET
-                new_file_name = rf'{pattern_result.groups()[0]}_{suffix_int}.nii.gz'
-                new_file_path = series_path.parent.joinpath(new_file_name)
-                return new_file_path
+                try:
+                    suffix_char = groups[1]
+                    suffix_int = ord(suffix_char) - self.CHAR_OFFSET
+                    new_file_name = rf'{groups[0]}_{suffix_int}.nii.gz'
+                    new_file_path = series_path.parent.joinpath(new_file_name)
+                    return new_file_path
+                except TypeError:
+                    print(groups)
+                    traceback.print_exc()
 
     def rename_file_only(self, series_path: pathlib.Path, pattern: re.Pattern):
         pattern_result = pattern.match(series_path.name)
@@ -61,7 +64,6 @@ class ProcessingStrategy(metaclass=ABCMeta):
             swan_pattern = self.pattern.match(series_path.name)
             if swan_pattern:
                 file_list.append(series_path)
-        print(file_list)
         for series_path in file_list:
             if series_path.exists():
                 if len(file_list) == 1:
@@ -72,7 +74,6 @@ class ProcessingStrategy(metaclass=ABCMeta):
                 file_base_name = series_path.name.replace('.nii.gz', '')
                 all_rename_file_list = list(filter(lambda x: x.stem == file_base_name, study_path.iterdir()))
                 if new_file_path:
-                    print(series_path, new_file_path)
                     new_file_base_name = new_file_path.name.replace('.nii.gz', '')
                     for rename_file in all_rename_file_list:
                         new_rename_file = new_file_path.parent.joinpath(f"{new_file_base_name}{rename_file.suffix}")
@@ -81,9 +82,10 @@ class ProcessingStrategy(metaclass=ABCMeta):
 
 
 class ADCProcessingStrategy(ProcessingStrategy):
-    pattern = re.compile(r'(?<!e)(ADC[a-z]{0,2}?)(\.nii\.gz)$', re.IGNORECASE)
-    suffix_pattern = re.compile(r'(?<!e)(ADC)([a-z]{0,2}?)(\.nii\.gz)$', re.IGNORECASE)
-    dwi_pattern = re.compile(r'(DWI0)(.*\.nii\.gz)$', re.IGNORECASE)
+    # pattern = re.compile(r'(?<!e)(ADC[a-z]{0,1}?)(\.nii\.gz)$', re.IGNORECASE)
+    pattern = re.compile(r'(?<!e)(ADC[a-z]{0,1}?)(\.nii\.gz)$', )
+    suffix_pattern = re.compile(r'(?<!e)(ADC)([a-z]{0,1}?)(\.nii\.gz)$',)
+    dwi_pattern = re.compile(r'(DWI0)([a-z]{0,1}?)(\.nii\.gz)$', re.IGNORECASE)
     FILE_SIZE = 100 * 1024  # 100kB
 
     def update_header(self, study_path: pathlib.Path, *args, **kwargs):
@@ -149,7 +151,6 @@ class ADCProcessingStrategy(ProcessingStrategy):
                         for dwi_file in dwi_file_list:
                             dwi_nii = nib.load(str(dwi_file))
                             if (dwi_nii.affine == adc_nii.affine).all():
-                                print(dwi_file)
                                 adc_file.unlink()
                                 adc_file_str = dwi_file.name.replace('DWI0', 'ADC')
                                 adc_file_path = adc_file.parent.joinpath(adc_file_str)
@@ -239,9 +240,10 @@ class T2ProcessingStrategy(ProcessingStrategy):
 
 
 class DwiProcessingStrategy(ProcessingStrategy):
-    pattern = re.compile(r'(DWI.*)(\.nii\.gz)$')
-    suffix_pattern = re.compile(r'(DWI.*)(?<![a-z])([a-z]{0,2}?)(\.nii\.gz)$')
-    FILE_SIZE = 550 * 1024  # 800kB
+    pattern = re.compile(r'(DWI0|DWI1000)(\.nii\.gz)$')
+    # suffix_pattern = re.compile(r'(DWI.*)(?<![a-z])([a-z]{0,2}?)(\.nii\.gz)$')
+    suffix_pattern = re.compile(r'(DWI0|DWI1000)([a-z]{0,2}?)(\.nii\.gz)$')
+    FILE_SIZE = 100 * 1024  # 800kB
 
     def process(self, study_path: pathlib.Path, *args, **kwargs):
         self.del_file(study_path=study_path)
@@ -255,8 +257,8 @@ class PostProcessManager:
                                                           T1ProcessingStrategy(),
                                                           T2ProcessingStrategy()]
 
-    def __init__(self, input_path: Union[str, pathlib.Path],*args, **kwargs):
-        self._input_path = pathlib.Path(input_path)
+    # def __init__(self, input_path: Union[str, pathlib.Path],*args, **kwargs):
+    #     self._input_path = pathlib.Path(input_path)
 
     def post_process(self, study_path):
         self.del_json_file(study_path=study_path)
@@ -266,25 +268,25 @@ class PostProcessManager:
     def del_json_file(self, study_path: pathlib.Path):
         json_path_list = filter(lambda x: x.name.endswith('json'), study_path.iterdir())
         for json_path in json_path_list:
-            json_path.unlink()
+            if json_path.exists():
+                json_path.unlink()
 
-    def run(self, executor: Union[ThreadPoolExecutor, None] = None):
-        is_dir_flag = all(list(map(lambda x: x.is_dir(), self.input_path.iterdir())))
-        if is_dir_flag:
-            study_path_list = list(self.input_path.iterdir())
-            for study_path in study_path_list:
-                self.post_process(study_path=study_path)
-        else:
-            self.post_process(study_path=self.input_path)
-            # break
-
-    @property
-    def input_path(self):
-        return self._input_path
-
-    @input_path.setter
-    def input_path(self, value: str):
-        self._input_path = pathlib.Path(value)
+    # def run(self, executor: Union[ThreadPoolExecutor, None] = None):
+    #     is_dir_flag = all(list(map(lambda x: x.is_dir(), self.input_path.iterdir())))
+    #     if is_dir_flag:
+    #         study_path_list = list(self.input_path.iterdir())
+    #         for study_path in study_path_list:
+    #             self.post_process(study_path=study_path)
+    #     else:
+    #         self.post_process(study_path=self.input_path)
+    #         # break
+    # @property
+    # def input_path(self):
+    #     return self._input_path
+    #
+    # @input_path.setter
+    # def input_path(self, value: str):
+    #     self._input_path = pathlib.Path(value)
 
 
 if __name__ == '__main__':
